+++ src/modules/tools/statisticDashboard/components/StatisticDashboard.vue
@@ -1,5 +1,6 @@
 <script>
 import {mapGetters, mapMutations} from "vuex";
+import TableComponent from "../../../../share-components/table/components/TableComponent.vue";
 import {getComponent} from "../../../../utils/getComponent";
 import ToolTemplate from "../../ToolTemplate.vue";
 import getters from "../store/gettersStatisticDashboard";
@@ -8,7 +9,24 @@ import mutations from "../store/mutationsStatisticDashboard";
 export default {
     name: "StatisticDashboard",
     components: {
-        ToolTemplate
+        ToolTemplate,
+        TableComponent
+    },
+    data () {
+        return {
+            data: {
+                headers: ["Raumeinheit", "2023", "2022"],
+                items: [
+                    ["Harburg", 1234, 1234],
+                    ["Ludwigslust Parchim", 23456, 1234],
+                    ["Lübeck", 23475, 1234],
+                    ["Niedersachsen", 34844, 1234]
+                ]
+            },
+            selectMode: "column",
+            showHeader: true,
+            sortable: true
+        };
     },
     computed: {
         ...mapGetters("Tools/StatisticDashboard", Object.keys(getters))
@@ -41,7 +59,12 @@ export default {
         :deactivate-gfi="deactivateGFI"
     >
         <template #toolBody>
-            test
+            <TableComponent
+                :data="data"
+                :select-mode="selectMode"
+                :show-header="showHeader"
+                :sortable="sortable"
+            />
         </template>
     </ToolTemplate>
 </template>